+++ crates/release_plz/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- add shellcompletions command
+
 ## [0.2.17] - 2022-05-29
 
 ### Added